:recycle: Move nonNullable function to root
